add spread gossip and listen gossip with discovery module
