+++ services-api/src/main/java/io/scalecube/services/discovery/api/ServiceDiscovery.java
@@ -2,6 +2,7 @@ package io.scalecube.services.discovery.api;
 
 import io.scalecube.services.ServiceEndpoint;
 import io.scalecube.services.ServiceLoaderUtil;
+import io.scalecube.services.api.ServiceMessage;
 import io.scalecube.transport.Address;
 
 import reactor.core.publisher.Flux;
@@ -9,6 +10,12 @@ import reactor.core.publisher.Mono;
 
 public interface ServiceDiscovery {
   public static final String SERVICE_METADATA = "service";
+  
+  Address address();
+
+  ServiceEndpoint endpoint();
+
+  Mono<String> spreadGossip(ServiceMessage gossip);
   
   static ServiceDiscovery getDiscovery() {
     ServiceDiscovery discovery = ServiceLoaderUtil.findFirstMatched(ServiceDiscovery.class)
@@ -19,11 +26,9 @@ public interface ServiceDiscovery {
 
   Mono<ServiceDiscovery> start(DiscoveryConfig discoveryConfig);
 
-  Mono<Void> shutdown();
-
-  Address address();
-
-  ServiceEndpoint endpoint();
+  Mono<Void> shutdown();  
   
   Flux<DiscoveryEvent> listen();
+
+  Flux<ServiceMessage> listenGossip();
 }
